An endpoint has been added to receive channels that the user is subscribed to.

+++ linksy_user/linksy_user/src/main/java/com/emil/linksy_user/model/PostResponse.java
@@ -17,6 +17,6 @@ public class PostResponse {
     private String voiceUrl;
     private String text;
     private String publishDate;
-    private int likesCount;
-    private int repostsCount;
+    private Long likesCount;
+    private Long repostsCount;
 }
